requirements.txt
pandas==3.0.2
numpy==2.4.4
matplotlib==3.10.8
seaborn==0.13.2
scikit-learn==1.8.0
scipy==1.17.1
jupyter==1.0.0
ipykernel==7.2.0
streamlit==1.28.1
Pillow==10.1.0
